refactor: Move getApolloServer to lib/utils refactor: Move createTestClient to tests/utils
