+++ lib/index.ts
@@ -14,4 +14,4 @@ export * from './services/gql-execution-context';
 export * from './tokens';
 export * from './type-factories';
 export * from './type-helpers';
-export * from './testing';
+export { getApolloServer } from './utils';
