--- lib/testing/index.ts
@@ -1,1 +1,0 @@
-export * from './create-apollo-test-client';
