+++ tests/e2e/code-first.spec.ts
@@ -1,7 +1,7 @@
 import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { ApolloServerTestClient } from 'apollo-server-testing';
-import { createTestClient } from '../../lib';
+import { createTestClient } from '../utils/create-test-client';
 import { ApplicationModule } from '../code-first/app.module';
 import { gql } from 'apollo-server-express';
